cli/bhej/requirements.txt
appdirs==1.4.4
apturl==0.5.2
arrow==0.15.7
asgiref==3.2.7
asn1crypto==0.24.0
attrs==19.3.0
binaryornot==0.4.4
black==19.10b0
blinker==1.4
boto3==1.16.7
botocore==1.19.7
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.3.7
command-not-found==0.3
cookiecutter==1.7.2
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.7.1
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.2
dnspython==1.16.0
docutils==0.14
entrypoints==0.3
flake8==3.7.8
Flask==1.1.2
gunicorn==20.0.4
httplib2==0.9.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
numpy==1.18.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pandas==1.0.3
pathspec==0.8.0
pexpect==4.2.1
Pillow==6.1.0
pkginfo==1.4.2
poyo==0.5.0
protobuf==3.0.0
psycopg2==2.8.5
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.15.0
python-http-client==3.2.7
python-magic==0.4.18
python-slugify==4.0.1
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
regex==2020.5.14
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
s3transfer==0.3.3
SecretStorage==2.3.1
selenium==3.141.0
sendgrid==6.3.0
sendgrid-django==4.2.0
sentry-sdk==0.11.2
simplejson==3.13.2
six==1.15.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
text-unidecode==1.3
toml==0.10.1
tqdm==4.23.4
twilio==6.40.0
typed-ast==1.4.1
typer==0.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.3
virtualenv==16.0.0
wadllib==1.3.2
Werkzeug==1.0.1
whitenoise==5.0.1
xkit==0.0.0
yapf==0.28.0
zope.interface==4.3.2
